[4596] Refactor string extensions for appending query parameters to Stream CDN hosted image string URLs to the client module and make use of Android's `Uri` class for appending query params instead of creating a separate function
